LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/aspect-ratio/quirks-mode-002-expected.txt


FAIL body height equals width assert_equals: expected 784 but got 584
